+++ src/monitor-control-service/azext_amcs/manual/_params.py
@@ -6,6 +6,7 @@
 # pylint: disable=too-many-lines
 # pylint: disable=too-many-statements
 
+
 from azure.cli.core.commands.parameters import (
     tags_type,
     resource_group_name_type,
@@ -24,7 +25,7 @@ from azext_amcs.action import (
     AddDataSourcesWindowsEventLogs,
     AddDataSourcesSyslog
 )
-
+from azext_amcs.manual._validators import validate_association_name_with_endpoint
 
 from azext_amcs.vendored_sdks.amcs.models import KnownDataFlowStreams, KnownPerfCounterDataSourceStreams, \
     KnownWindowsEventLogDataSourceStreams, KnownSyslogDataSourceStreams, \
@@ -55,6 +56,8 @@ def load_arguments(self, _):
         c.argument('association_name', options_list=['--name', '-n'], help='The name of the association.')
         c.argument('description', help='Description of the association.')
         c.argument('rule_id', help='The resource ID of the data collection rule that is to be associated.')
+        c.argument('endpoint_id', help='The resource ID of the data collection endpoint that is to be associated.',
+                   validator=validate_association_name_with_endpoint)
         c.argument('data_collection_rule_name', options_list=['--rule-name'])
 
     with self.argument_context('monitor data-collection rule association list') as c: