+++ src/aks-preview/HISTORY.rst
@@ -13,10 +13,14 @@ Pending
 +++++++
 * Vendor new SDK and bump API version to 2023-09-02-preview.
 
-0.5.168
+0.5.169
 +++++++
 * Add `--node-soak-duration` to the `az aks nodepool add/update/upgrade` commands.
 * Add `--drain-timeout` to the `az aks nodepool add/update/upgrade` commands (already in [azure-cli](https://github.com/Azure/azure-cli/pull/27475)).
+
+0.5.168
++++++++
+* Add `--enable-image-integrity` to the `az aks update` command.
 
 0.5.167
 +++++++