+++ src/alb/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+1.0.0
+++++++
+* Upgrade API version to 2023-11-01 for GA.
+
 0.1.0
 ++++++
 * Initial release for ALB extension.